Fix partners review recommendations
