+++ src/modules/commentModal/postComments.js
@@ -1,5 +1,7 @@
+import { baseUrl, appId } from '../api.js';
+
 const postComments = async (id, username, comment, creationDate) => {
-  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/yOdv0JXV8ssnN9f9thVW/comments', {
+  const response = await fetch(`${baseUrl}/${appId}/comments`, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
